+++ routes/users.js
@@ -1,25 +1,8 @@
 var express = require('express');
 var router = express.Router();
 const bcrypt = require('bcryptjs');
-
-//추가한 부분
-var mysql = require('mysql');
-// Connection 객체 생성 
-var connection = mysql.createConnection({
-  host: 'localhost',
-  port: 3306,
-  user: 'root',
-  password: '1008',
-  database: 'gain_muscle'
-});
-// Connect
-connection.connect(function (err) {
-  if (err) {
-    console.error('mysql connection error');
-    console.error(err);
-    throw err;
-  }
-});
+var mysqlConnection = require('../utils/mysqlConnection');
+var connection = mysqlConnection.getConnection()
 
 router.get('/', function (req, res) {
   connection.query('SELECT * FROM users', function (err, rows) {
